+++ engine/src/test/java/net/jqwik/api/domains/ContextWithProviderMethods.java
@@ -4,6 +4,7 @@ import java.time.*;
 import java.util.*;
 
 import net.jqwik.api.*;
+import net.jqwik.api.Tuple.*;
 import net.jqwik.api.arbitraries.*;
 import net.jqwik.api.providers.*;
 
@@ -21,19 +22,34 @@ class ContextWithProviderMethods extends DomainContextBase {
 	}
 
 	@Provide
-	ListArbitrary<?> listsOfSize3(TypeUsage targetType) {
+	Arbitrary<List<String>> listsOfSize3(TypeUsage targetType) {
 		TypeUsage innerTarget = targetType.getTypeArgument(0);
-		try {
-			Arbitrary<Object> inner = Arbitraries.defaultFor(innerTarget);
+		if (innerTarget.isOfType(String.class)) {
+			Arbitrary<String> inner = Arbitraries.defaultFor(innerTarget);
 			return inner.list().ofSize(3);
-		} catch (CannotFindArbitraryException cannotFindArbitraryException) {
-			return null;
 		}
+		return null;
 	}
 
 	@Provide
 	Arbitrary<List<LocalDate>> listsOfDates() {
 		return Arbitraries.just(LocalDate.now()).list().ofSize(1);
+	}
+
+	@Provide
+	Arbitrary<Tuple2<Integer, Integer>> tuple2OfInts() {
+		return Arbitraries.integers().tuple2();
+	}
+
+	// This should not be applied to target type Tuple2<Integer, Integer>
+	@Provide
+	Arbitrary<Tuple2<String, String>> tuple2OfStrings() {
+		return Arbitraries.strings().tuple2();
+	}
+
+	@Provide
+	Arbitrary<Tuple3<Integer, Integer, Integer>> tuple3OfInts() {
+		return Arbitraries.integers().tuple3();
 	}
 
 	//@Provide